pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lora-attention-analyzer"
version = "0.1.0"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
description = "A tool for analyzing attention maps in Stable Diffusion models with LoRA support"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research", 
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Graphics",
]
keywords = ["stable-diffusion", "lora", "attention", "visualization", "ai"]

dependencies = [
    "torch>=1.13.0",
    "torchvision>=0.14.0", 
    "diffusers>=0.21.0",
    "transformers>=4.25.0",
    "daam",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "numpy>=1.21.0",
    "Pillow>=8.0.0",
    "safetensors>=0.3.0",
    "xformers>=0.0.16",
    "fire>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "flake8", 
    "isort",
]
optional = [
    "scikit-image>=0.19.0",
    "scipy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/lora-attention-analyzer"
"Bug Reports" = "https://github.com/yourusername/lora-attention-analyzer/issues"
"Source" = "https://github.com/yourusername/lora-attention-analyzer"

[project.scripts]
lora-attention-analyzer = "lora_attention_analyzer.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["lora_attention_analyzer*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88